Add StFinderSearchOptions class.
This change adds the StFinderSearchOptions class, which is a widget that encapsulates
the search options used in the Finder application (for now only regexp search). With this
change the StFinderSearchBar widget was cleaned up and the help button was moved
to StFinderPresenter.
A slight change of layout was also done, which makes the interface more compact.